requirements.txt
accounts==0.2.4
alembic==1.13.0
amqp==2.6.1
anyio==3.7.1
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
bcrypt==4.1.1
billiard==3.6.4.0
cachetools==5.3.2
celery==4.4.7
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==7.1.2
colorama==0.4.6
cryptography==41.0.7
cssselect==1.2.0
cssutils==2.9.0
Django==4.2.7
django-js-asset==2.1.0
django-mptt==0.15.0
django-paypal==2.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==1.3.1
emails==0.5.15
fastapi==0.70.1
greenlet==3.0.2
gunicorn==20.1.0
h11==0.9.0
idna==3.6
inflect==5.6.2
Jinja2==2.11.3
kombu==4.6.11
lxml==4.9.3
Mako==1.3.0
MarkupSafe==2.1.3
more-itertools==10.1.0
packaging==23.2
passlib==1.7.4
paypal==1.2.5
Pillow==10.1.0
pluggy==0.13.1
premailer==3.10.0
psycopg2-binary==2.9.9
py==1.11.0
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.13
pytest==5.4.3
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2023.3.post1
raven==6.10.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
sqlparse==0.4.4
starlette==0.16.0
tenacity==6.3.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.11.8
vine==1.3.0
wcwidth==0.2.12
websockets==8.1